stage 8 - iteration

+++ new_game/src/engine/console/console_globals.cpp
@@ -1,4 +1,3 @@
-#include <windows.h>
 #include "engine/console/console_globals.h"
 
 // uc_orig: CONSOLE_mess (fallen/DDEngine/Source/console.cpp)
@@ -11,8 +10,8 @@ ConsoleText console_Data[CONSOLE_LINES];
 CBYTE console_status_text[_MAX_PATH];
 
 // uc_orig: last_tick (fallen/DDEngine/Source/console.cpp)
-// claude-ai: BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD
-DWORD console_last_tick;
+// BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD → uint64_t
+uint64_t console_last_tick;
 // uc_orig: this_tick (fallen/DDEngine/Source/console.cpp)
-// claude-ai: BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD
-DWORD console_this_tick;
+// BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD → uint64_t
+uint64_t console_this_tick;
