+++ new_game/src/game/game_globals.cpp
@@ -70,8 +70,8 @@ SLONG bullet_counter = 0;
 UBYTE screen_mem[640 * 3][480];
 
 // uc_orig: already_warned_about_leaving_map (fallen/Source/Game.cpp)
-// claude-ai: BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD
-DWORD already_warned_about_leaving_map = 0;
+// BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD → uint64_t
+uint64_t already_warned_about_leaving_map = 0;
 
 // uc_orig: draw_map_screen (fallen/Source/Game.cpp)
 UBYTE draw_map_screen = 0;
