+++ new_game/src/things/characters/person_globals.cpp
@@ -162,11 +162,11 @@ SLONG stat_arrested_innocent;
 // uc_orig: stat_count_bonus (fallen/Source/Person.cpp)
 SLONG stat_count_bonus;
 // uc_orig: stat_start_time (fallen/Source/Person.cpp)
-// claude-ai: BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD
-DWORD stat_start_time;
+// BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD → uint64_t
+uint64_t stat_start_time;
 // uc_orig: stat_game_time (fallen/Source/Person.cpp)
-// claude-ai: BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD
-DWORD stat_game_time;
+// BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD → uint64_t
+uint64_t stat_game_time;
 
 // uc_orig: player_visited (fallen/Source/Person.cpp)
 UBYTE player_visited[16][128];
