+++ new_game/src/ui/frontend/attract.cpp
@@ -197,8 +197,8 @@ void ScoresDraw(void)
     extern SLONG stat_arrested_thug;
     extern SLONG stat_arrested_innocent;
     extern SLONG stat_count_bonus;
-    // claude-ai: BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD
-    extern DWORD stat_start_time, stat_game_time;
+    // BUGFIX-OC-TICK-OVERFLOW: SLONG → DWORD → uint64_t
+    extern uint64_t stat_start_time, stat_game_time;
 
     SLONG count = 0, count_bonus_left = 0, c0;
     SLONG ticks, h, m, s;
